app/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
alembic
python-dotenv
pydantic[email]
pyjwt
bcrypt==4.0.1
passlib[bcrypt]
python-multipart